# docs/swagger.yaml
basePath: /api
definitions:
...
  BanResponse:
    properties:
      id:
        type: string
      image:
        type: string
      username:
        type: string
    type: object
...
  ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/HttpError'
    type: object
  ErrorsResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/FieldError'
        type: array
    type: object
  FieldError:
    properties:
      field:
        description: The property containing the error
        type: string
      message:
        description: The specific error message
        type: string
    type: object
...
  HttpError:
    properties:
      message:
        description: The specific error message
        type: string
      type:
        description: The Http Response as a string
        type: string
    type: object
...
  MemberRequest:
    properties:
      memberId:
        type: string
    type: object
  MemberSettings:
    properties:
      color:
        type: string
      nickname:
        type: string
    type: object
  MemberSettingsRequest:
    properties:
      color:
        type: string
      nickname:
        type: string
    type: object
...
  SuccessResponse:
    properties:
      success:
        description: Only returns true, not a json object
        type: boolean
    type: object
...
host: localhost:<PORT>
info:
  contact: {}
  description: Xonia REST API . This service uses sessions for authentication
  license:
    name: GNUv3
  title: Xonia API
  version: "2.0.0"
paths:
...
  /guilds/{guildId}/bans:
    delete:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      - description: Member ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/SuccessResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Unban Member
      tags:
      - Members
    get:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/BanResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get Guild Ban list
      tags:
      - Members
    post:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      - description: Member ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/SuccessResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Ban Member
      tags:
      - Members
...
  /guilds/{guildId}/kick:
    post:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      - description: Member ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/SuccessResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Kick Member
      tags:
      - Members
  /guilds/{guildId}/member:
    get:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MemberSettings'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get Member Settings
      tags:
      - Members
    put:
      parameters:
      - description: Edit Member
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MemberSettingsRequest'
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Edit Member Settings
      tags:
      - Members
...
swagger: "2.0"